changes address module

See merge request global-design/whybuy!93

+++ app/Http/Controllers/Api/UserController.php
@@ -153,7 +153,13 @@ class UserController extends Controller
         try {
             $user = auth()->user();
             if(!empty($user)){
-                $addresses = $this->addresses->where(['id' => $request->id,'user_id' => $user->id])->first();
+                $validator= Validator::make($request->all(),[
+                    'address_id'=>'required|integer|exists:addresses,id',
+                ]);
+                if($validator->fails()){
+                    return $this->ErrorResponse(400,$validator->errors()->first());
+                }
+                $addresses = $this->addresses->where(['id' => $request->address_id,'user_id' => $user->id])->first();
                 if(!empty($addresses)){
                     return $this->SuccessResponse(200, 'Address get successfully',$addresses);
                 }
@@ -173,17 +179,18 @@ class UserController extends Controller
         try {
             $user = auth()->user();
             if(!empty($user)){
-                $addresses = $this->addresses->where(['id' => $request->id,'user_id' => $user->id])->first();
+                $validator= Validator::make($request->all(),[
+                    'address_id'=>'required|integer|exists:addresses,id',
+                    'address'=>'required',
+                    'city'=>'required',
+                    'state'=>'required',
+                    'country_id'=>'required|exists:countries,id',
+                ]);
+                if($validator->fails()){
+                    return $this->ErrorResponse(200,$validator->errors()->first());
+                }
+                $addresses = $this->addresses->where(['id' => $request->address_id,'user_id' => $user->id])->first();
                 if(!empty($addresses)){
-                    $validator= Validator::make($request->all(),[
-                        'address'=>'required',
-                        'city'=>'required',
-                        'state'=>'required',
-                        'country_id'=>'required|exists:countries,id',
-                    ]);
-                    if($validator->fails()){
-                        return $this->ErrorResponse(200,$validator->errors()->first());
-                    }
                     $addresses->address = $request->address;
                     $addresses->city = $request->city;
                     $addresses->state = $request->state;
@@ -193,7 +200,7 @@ class UserController extends Controller
                     $addresses->save();
                     return $this->SuccessResponse(200, 'Address updated successfully',$addresses);
                 }
-            return $this->ErrorResponse(200, 'Address not found');
+                return $this->ErrorResponse(200, 'Address not found');
             }
             return $this->ErrorResponse(200, 'Something Went Wrong');
         } catch (Exception $exception) {
@@ -209,7 +216,14 @@ class UserController extends Controller
         try {
             $user = auth()->user();
             if(!empty($user)){
-                $addresses = $this->addresses->where(['id' => $request['id'],'user_id' => $user->id])->first();
+                $validator= Validator::make($request->all(),[
+                    'address_id'=>'required|integer|exists:addresses,id',
+                ]);
+                if($validator->fails()){
+                    return $this->ErrorResponse(200,$validator->errors()->first());
+                }
+
+                $addresses = $this->addresses->where(['id' => $request->address_id,'user_id' => $user->id])->first();
                 if(!empty($addresses)){
                     $addresses->delete();
                     return $this->SuccessResponse(200, 'Address deleted successfully');
